api/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
numpy
requests
tensorflow==2.11.0
pandas
feather-format
scikit-learn
joblib 